finnaly version oop
