+++ src/index.html
@@ -42,9 +42,9 @@
         </button>
         <form id="form-profile" name="form-profile" class=" form popup__form form__set" id="form-profile">
           <h3 class="popup__heading">Редактировать профиль</h3>
-          <input id="name-input" class="popup__input name_input form-input" type="text" value="Жак-ив Кусто" name="yourname" minlength="2" maxlength="40" pattern="^[a-zA-Zа-яёА-ЯЁ\-\s]+" required placeholder="Ваше имя" data-error-message="Разрешены только латинские, кириллические буквы, знаки дефиса и пробелы">
+          <input id="name-input" class="popup__input name_input form-input" type="text" value="" name="yourname" minlength="2" maxlength="40" pattern="^[a-zA-Zа-яёА-ЯЁ\-\s]+" required placeholder="Ваше имя" data-error-message="Разрешены только латинские, кириллические буквы, знаки дефиса и пробелы">
           <span class="form__input-error name-input-error"></span>
-          <input id="job-input" class="popup__input job_input form-input" type="text" value="Исследователь окена" name="aboutyou" minlength="2" maxlength="200" pattern="^[a-zA-Zа-яёА-ЯЁ\-  \s]+$" required placeholder="Ваша работа" data-error-message="Разрешены только латинские, кириллические буквы, знаки дефиса и пробелы">
+          <input id="job-input" class="popup__input job_input form-input" type="text" value="" name="aboutyou" minlength="2" maxlength="200" pattern="^[a-zA-Zа-яёА-ЯЁ\-  \s]+$" required placeholder="Ваша работа" data-error-message="Разрешены только латинские, кириллические буквы, знаки дефиса и пробелы">
           <span class="form__input-error job-input-error"></span>
           <button type="submit" data-value="Сохранение..." value="Сохранить" class="popup__button form__submit">Сохранить</button>
         </form>
